Debug Level 7 Save
Added validateStringFilenameUsingIO
