+++ src/main/java/formatting/Helper.java
@@ -1,10 +1,14 @@
 package formatting;
 
+import java.io.File;
 import java.time.format.DateTimeFormatter;
+import java.util.Scanner;
+
 public class Helper {
     //helper class for formatting
 
     public static DateTimeFormatter formatter = DateTimeFormatter.ofPattern ("d/MM/yyyy HHmm");
+    public static String dukeFilePath = System.getProperty("user.home") + File.separator + "Desktop" + File.separator + "dukeFile";
 
     public static void newline(){
         System.out.println("\n");
@@ -13,6 +17,8 @@ public class Helper {
         System.out.println("____________________________________");
     }
 
-
-
+    public static String userInput() {
+        Scanner in = new Scanner(System.in);
+        return in.nextLine();
+    }
 }
